requirements.txt
Flask
PyYAML
pyaudio
openai
requests
